src/main/webapp/app/entities/books-option-details/delete/books-option-details-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { BooksOptionDetailsService } from '../service/books-option-details.service';

import { BooksOptionDetailsDeleteDialogComponent } from './books-option-details-delete-dialog.component';

describe('BooksOptionDetails Management Delete Component', () => {
  let comp: BooksOptionDetailsDeleteDialogComponent;
  let fixture: ComponentFixture<BooksOptionDetailsDeleteDialogComponent>;
  let service: BooksOptionDetailsService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [BooksOptionDetailsDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(BooksOptionDetailsDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(BooksOptionDetailsDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(BooksOptionDetailsService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
